Add files via upload
Approximately 3.5 billion searches are performed on Google daily, which means that approximately 40,000 searches are performed every second on Google. So Google search is a great use case for analyzing data based on search queries. Google doesn’t give much access to the data about daily search queries, but another application of google known as Google Trends can be used for the task of Google search analysis. Google Trends provides an API that can be used to analyze the daily searches on Google. This API is known as pytrends.

Here I will be analyzing the Google search trends on the queries based on “Data analyst”, so let’s create a DataFrame of the top 10 countries which search for “Data analyst” on Google.